FrontEnd/renewable-energy-dashboard/requirements.txt
# Microgrid System Backend Dependencies
# Compatible with Python 3.8+ (including Python 3.13)

# FastAPI and ASGI server
fastapi>=0.100.0,<0.110.0
uvicorn[standard]>=0.20.0,<0.25.0

# MQTT client
paho-mqtt>=1.6.0,<2.0.0

# Data validation and serialization (compatible versions)
pydantic>=2.0.0,<2.6.0
pydantic-core>=2.0.0,<2.15.0

# HTTP client for potential external API calls
httpx>=0.24.0,<0.26.0

# Date and time utilities
python-dateutil>=2.8.0,<3.0.0

# Environment variables management
python-dotenv>=1.0.0,<2.0.0

# Logging and monitoring
structlog>=23.0.0,<24.0.0

# Additional dependencies for better compatibility
typing-extensions>=4.5.0
annotated-types>=0.4.0

# Optional: For enhanced MQTT features
# paho-mqtt[ssl]>=1.6.0,<2.0.0  # Uncomment if you need SSL/TLS support
